change to unsigned int
(cherry picked from commit 9036a16a77ba75453dbf9ba97fa2a126adf80277)

+++ backend/src/main/java/top/fengpingtech/solen/bean/ConnectionBean.java
@@ -15,8 +15,8 @@ public class ConnectionBean {
     private String deviceId;
     private String serverHost;
     private String serverPort;
-    private Integer lac;
-    private Integer ci;
+    private Long lac;
+    private Long ci;
     private ConnectionStatus status;
     private Integer header;
     private Integer inputStat;
